overcloud-source: Fix overcloud parameters.
These values were missed in the initial overcloud-source.yaml.

nova.metadata-proxy is required to allow vms to communicate with the
EC2-style metadata service.

quantum.ovs.fixed_range constraints DHCP allocation for the undercloud,
but in the overcloud we do not want this constraint because the individual
subnets we define will do that. quantum.ovs.ovs_range configures the
range of ips that can be assigned to the tenant subnets.

Change-Id: If7191fa8baf1209ae33b9e8200f52ea0ade97810