Configure undercloud to use ext4 ephemeral FS
Ext4 is significantly more efficient when being created, and we do not
need to support older OS's or platforms that do not support ext4.

This will be ineffective until the change to the nova element lands,
but it is harmless to land this first. That change is referenced by this
Gerrit change ID in tripleo-image-elements:

Ic754f3b342b632a6919d4ad1b808627ec9fa4308

Closes-Bug: #1266262
Change-Id: Icca5fd94ca230141818ee9adb416f6f062869c6b