Make it possible to do physical networks for VMs.
This patch makes it possible to setup physical networks for VMs,
separate to the control plane configuration which is needed for
routed/natted access to physical networks.

Future work is needed to automate ci-overcloud configuration of the
control plane where we need two distinct bridges, but this is enough
to stop folk dying of boredom setting up a sizeable ci-overcloud.

Change-Id: I6ac7129f22bb797467adb0408638781d20081f19