Adds block-storage-nfs template
To support the changes to the cinder element to allow the nfs
backend to be utilized, this template has been added to show
the usage of the nfs_shares key in the cinder metadata.
The value is a list of strings containing share addresses.

This change is added to facilitate an example to test
the cinder element change: https://review.openstack.org/#/c/74563/

You may setup the nfs server manually or use the nfs-server
element at https://review.openstack.org/#/c/74712/.

Change-Id: I5b6cb118b34421ea07a81ed1fe68db24b1d4f19d