Make Wait Conditions depend on config creation
Currently our wait conditions are racing with Heat resolving
configurations. There should be plenty of time but sometimes Heat may
be dealing with a temporarily problematic Nova API and spinning on that.
While that is happening, the in-instance tools will not have their full
configuration available to them. We don't want the wait condition timeout
to start until the box has had its actual config exposed to it.

Change-Id: I0eab8fe7547d3cbcebb1559fd3d06206b1750e96