Merge Block Storage into Overcloud
This patch removes the seperate overcloud-with-block-storage.yaml
make target and merges the functionality into the main overcloud.yaml
target.  The default BLOCKSTORAGESCALE is 0. Change this number to add
additional block storage nodes.

This patch also updates the overcloud-with-block-storage-nfs target
to more closely resemble the overcloud.yaml target, including swift
and ssl templates as needed. The default BLOCKSTORAGESCALE is 1.
Change this number to add additional block storage nodes.

Change-Id: Ifcc733cfbf9f356393c3f6c72b2bbbb0214d3040