Expose dnsmasq options
Adds a new parameter, NeutronDnsmasqOptions, to the overcloud template.

Allows the ability to set dnsmasq options for neutron dhcp agent. This
will allow us to configure mtu to be 1400 for tenant instances on the
overcloud.  This should help with poor network performance and vm's that
are just plain unreachable via ssh due to the GRE tunnel overhead.

The default here has been set to:
dhcp-option-force=26,1400

This is the recommended way to configure OpenStack with the Open vSwitch
plugin per:
http://docs.openstack.org/admin-guide-cloud/content/openvswitch_plugin.html

All the documentation I can find on the web (openstack-dev,
ask.openstack.org, etc), recommend applying this setting. Others have
reported slow vm performance as well, and this resolves that issue
(apparently anyway...we'd need to test).

Change-Id: If24326045987b5a484ba2f71f591092987966536
Partial-Bug: #1270646