Add BlockStorage0Config Resource
Add a BlockStorage0Config resource in the block-storage.yaml template.
This was missing previously and was causing the behavior of having all
block storage nodes only being created after all other nodes were up in
a deployment (I presume b/c of the wait conditions).

Also, we need the BlockStorage0Config resource so that we can get the
configuration of neutron-openvswitch-agent from the metadata, previously
the values in the file were unset, causing the service to fail to start.

Change-Id: I297de24079d1ece66d35213b4ef1f2c8c50e73f0