Add mysql innodb buffer pool size
Innodb buffer pool size in mysql config file is now set to 592M,
this value is for VMs with 2G RAM too much. This patch makes the
pool size configurable and decreases default value to 100 MB
(this should be sufficient for default devtest setup). A patch which
for tripleo-image-element which sets this value in my.cnf will follow.

Change-Id: I0d2529abe51d4dc34415bf6c40da7a267bf3e65b