Set glance protocol and port as Heat configurable
This change is to set the glance protocol and port as
configurable via the Heat template. Presently the port is
hard-coded in the elements nova.conf file, and the protocol
is assumed as being the default (http).

This change will allow the glance_api_servers
to be set (in nova.conf) using the constituent parts:
    glance_protocol://glance_host:glance_port

Change to nova.conf to read this value is:
    Idccc0d60c9f6b17a853c6de1bbea64bfc7e028b2

Default port value is set to the nova default(9292) which is
currently hard-coded in the elements nova.conf file.
Default protocol value is set to the nova default(http).

Change-Id: I3c7218292797c62c36e2aaab4f325bf053ef140b