Increase PXE deployment timeout for overcloud
The PXE deployment often times out on baremetal deployment with >5
overcloud nodes because the time being measured includes the
dd of the image, which can be slow when many images are sent
from the same undercloud host.

Ideally we'd make the image sizes smaller, and/or make the
undercloud cache the images more sanely.  It would also be
possible to split the timout to measure dd-time and boot-time
separately, but for now we just make the timeout configurable so
a user can raise it if they have problems.

Change-Id: I15540eec7a68eab4c9d128b65a95b1c0a2b64582
Co-Author: nicholas.randon@hp.com