Parameterize network type and tunnel types
Remove the hardcoding of gre as the Neutron tenant network type for the
Overcloud. This will enable the ability to deploy an Overcloud that uses
vxlan instead of gre tunnels. A new parameter, NeutronTunnelTypes, is
added to allow configuring the tunnel_types parameter in the Neutron ML2
configuration.

This change is required by https://review.openstack.org/#/c/92913

Change-Id: I2c2e2153a61349e58ada28c87aa2338c9f00e7bd