Undercloud params for compute driver and manager
Add undercloud template parameters for NovaComputeDriver,
NovaComputeManager, and NovaSchedulerHostManager.

The motivation here is to be able to test configure Nova to use
the new (in-tree) Ironic compute driver (which used to live
in Ironic).

NOTE: The initial Nova in-tree Ironic driver won't have its
own ClusteredComputeManager (compute_manager). Even so I've
gone ahead and added a parameter for NovaComputeManager
so we can modify this accordingly in the future without
making hard coded template changes.

Change-Id: Ib48a6b6d8a6cff157bdf3948bd0330e9b29dd46a