Add SwiftMinPartHours parameter
This parameter adds the facility for an operator to set the length
of time in hours that a partition in a swift ring can be altered
following a rebalance of that ring. The default value is set to one.

The corresponding element change is here:
Ie6438386a54794e1ae2f31ad925db01c2c10ed6d

Change-Id: Icc5e431a7e2884b3ca3a255b6fd901619bc98460