Add swift.mount-check metadata
We've submitted a patch (https://review.openstack.org/#/c/130172/)
to set the value of mount_check to swift.mount-check if it exists,
and otherwise to set mount_check to false.  By default TripleO
deployments set mount_check to false since they do not use mounted
disks to store data.  However we (HP) and others are now using
TripleO to deploy Swift servers with mounted drives for data, in
which case mount_check should be set to True.  This change adds
swift.mount-check data and sets it to the value of the
SwiftMountCheck parameter, which has a default value of False.

Change-Id: I36fece56bafa9fe9c4883b572687b3fc819eeae1