bump up the number of sesson limit for rabbitmq
the default maxconn is only 150, which maybe good for api services
but not enough for the rabbitmq session in a cluster as small as 15
nodes. so bump up the number to 1500 for rabbitmq to allow for 100
nodes. this number should be calculated based on the scale numbers
in the long run.
Closes-bug: #1386406

Change-Id: Ieb707b31022a6fc9ade32ed2a332b67bf4dc0311