Puppet: Ceilometer controller support
This patch adds support for the Ceilometer controller
role including the Ceilometer:

 -API
 -central agent
 -alarm notifier
 -alarm evaluator
 -collector
 -expirer

In order to enable swift metering the swift::proxy ceilometer middleware
was added in.

Also, a minor adjustment to the existing ceilometer HA proxy setting
was made to accommodate ceilometer auth settings. (not exactly sure
why but this seems to be required)

Like upstream TripleO Ceilometer is currently using a MySQL database
backend. A follow on patch can support configuring MongoDB for use
with Ceilometer.

Change-Id: I4e171274bd7679d386d93492d13dfa7c5d37f6a8