Deprecate the old mergepy overcloud templates
This patch moves all the related mergepy templates for the
overcloud into a deprecated directory. The Makefile has
been updated so that overcloud.yaml is still generated
at the top level so this shouldn't break end users.

This is to reduce confusion for new users who are learning the
TripleO heat templates and find the fact that we have two full
implementations very confusing.

Change-Id: I0848aca4dee3e37cb4c6089c5f655ad22ac6c5fd