Make all default values match overcloud defaults
It's very confusing for them to be different, especially in the case of
comparing Tuskar vs non-Tuskar deployments where the parameters are read
from different files.

Note: NeutronPhysicalBridge is named differently in the overcloud
template (HypervisorNeutronPhysicalBridge). This is the only parameter
checked that isn't named exactly the same, hopefully there aren't any
others.

(Checked controller, compute, ceph, cinder, and swift for both puppet
and non-puppet templates)

Change-Id: I48ce1eb40d2d080c589ce619c50eddff17efe882