Add network templates for multiple NIC configuration
This change adds a set of network interface configurations for use
with network isolation. The multiple-nics templates includes one
separate NIC per network, and assumes that nic1 is used for the
provisioning network (ctlplane). Also included is an environment
file for including the multiple-nics configuration in a deployment.

This revision changes the ordering of the NICs. By doing that, it
is possible to wire up only a subset of the NICs for the storage
nodes, and it is possilbe to leave the External NIC only configured
on the controllers.

rdo: Updated this commit for static control plane configuration

gfidente: Includes Iabd835724848d754d5522968e1c8e3cf9f78e6c6 to
pass pep8 testing.

Co-Authored-By: Rhys Oxenham <roxenham@redhat.com>
Depends-On: Ie21282eb8cc0a8767aceb270bb1c892cec51bf73
Depends-On: Ia2a10318de47854a9b8932d72243ce0a85603556
Depends-On: I5126887ff94f7a48d6b3160353b802539491de04
Change-Id: Ic878d1ed1a85b5705295d087a743570ca8213504
(cherry picked from commit 6c57eab4c29cca7a80c13eae83b6126513621e4d)