Allow a user to set specific setting per host
It is currently not possible to specify settings per host and not per
type of host.

One of the example of the problematic that could cause is : What if
node0 have devices /dev/sdb and /dev/sdc while node1 have devices
/dev/sda and /dev/sdd, they is currently no way to specify that simply.

The idea here is to add a top priority file in the hiera lookup that
will match the UUID of the System Information section in the output of
the dmidecode command.

The file could be provided with the firstboot/rsync stack for example.

Change-Id: I3ab082c8ebd2567bd1d914fc0b924e19b1eff7d0