Remove nameserver lookup dependency from Cisco ML2 CollectMacConfig.
The collection of hostname to MAC mappings done in AllNodesPostDeploy
uses 'hostname -f' to get the FQDN for each node.  This form
of the command causes a nameserver lookup for the domain name.  A
timing issue has been seen where the hostname lookup fails due to
the nameserver not having the mapping yet.  The solution is to
hardcode the domain to 'localdomain' as is done in a few other
patches--ie. see controller-puppet.yaml.

Change-Id: Ibea50fcc6b9f22ca163ff063e0dc9ca69dff5f34