Inject TLS certificate and keys for the Overcloud
This is a first implementation of adding TLS termination to the load
balancer in the controllers. The implementation was made so that the
appropriate certificate/private key in PEM format is copied to the
appropriate controller(s) via a software deployment resource.

And the path is then referenced on the HAProxy configuration, but this
part was left commented out because we need to be able to configure the
keystone endpoints in order for this to work properly.

Conflict:
    puppet/controller.yaml

(Conflict due to not backporting I0e86318eb5521e540cc567ce9d77e1060086d48b )

Change-Id: I0ba8e38d75a0c628d8132a66dc25a30fc5183c79
(cherry picked from commit 97b12afbadeadac0be348b7cc263b090f6e6f0b8)