Pass ceph::pool arguments when calling class
Pass the ceph::pool properties as arguments to the class call
instead of setting them as class defaults.

Ceph recommends max 32 PGs and min 4 PGs per OSD so this change
also lowers the defaults to 32 which works with 1 OSD, suits well
a scenario with 3 OSDs and is easy to customize in the static
hiera if more than 8 OSDs are deployed.

More info at: https://bugzilla.redhat.com/show_bug.cgi?id=1252546

Change-Id: Ifed11d1857900b2251dfdf69d6b6f168150e6330